feat(unification): remove base configuration, place native into original #370
