+++ PrebidMobile/PrebidMobile-core/src/test/java/org/prebid/mobile/ResultCodeTest.java
@@ -30,7 +30,6 @@ import org.prebid.mobile.tasksmanager.BackgroundThreadExecutor;
 import org.prebid.mobile.tasksmanager.TasksManager;
 import org.prebid.mobile.testutils.BaseSetup;
 import org.prebid.mobile.testutils.MockPrebidServerResponses;
-import org.prebid.mobile.units.configuration.AdUnitConfiguration;
 import org.robolectric.Robolectric;
 import org.robolectric.RobolectricTestRunner;
 import org.robolectric.Shadows;
