+++ PrebidMobile/PrebidMobile-core/src/test/java/org/prebid/mobile/rendering/networking/urlBuilder/AutoDetectedOpenRtbTest.java
@@ -29,7 +29,6 @@ import org.prebid.mobile.rendering.networking.parameters.*;
 import org.prebid.mobile.rendering.sdk.ManagersResolver;
 import org.prebid.mobile.rendering.utils.helpers.AdIdManager;
 import org.prebid.mobile.rendering.utils.helpers.AppInfoManager;
-import org.prebid.mobile.units.configuration.AdUnitConfiguration;
 import org.robolectric.Robolectric;
 import org.robolectric.RobolectricTestRunner;
 import org.robolectric.annotation.Config;
