+++ PrebidMobile/PrebidMobile-core/src/test/java/org/prebid/mobile/rendering/video/VideoDownloadTaskTest.java
@@ -29,7 +29,6 @@ import org.mockito.Mockito;
 import org.prebid.mobile.rendering.loading.FileDownloadListener;
 import org.prebid.mobile.rendering.networking.BaseNetworkTask;
 import org.prebid.mobile.test.utils.ResourceUtils;
-import org.prebid.mobile.units.configuration.AdUnitConfiguration;
 import org.robolectric.Robolectric;
 import org.robolectric.RobolectricTestRunner;
