+++ PrebidMobile/PrebidMobile-core/src/test/java/org/prebid/mobile/rendering/views/base/BaseAdViewTest.java
@@ -26,7 +26,6 @@ import org.mockito.Mockito;
 import org.prebid.mobile.rendering.errors.AdException;
 import org.prebid.mobile.rendering.views.AdViewManager;
 import org.prebid.mobile.test.utils.WhiteBox;
-import org.prebid.mobile.units.configuration.AdUnitConfiguration;
 import org.robolectric.Robolectric;
 import org.robolectric.RobolectricTestRunner;
 import org.robolectric.annotation.Config;
